requirement.txt
facenet-pytorch
tqdm
albumentations
brevitas
onnxoptimizer
protobuf==3.19.1
joblib
